Move all the files related to the OpenSearch service to the existing OpenSearch folder
