+++ src/django/api/services/opensearch/search.py
@@ -1,6 +1,6 @@
 import logging
-from api.services.opensearch import OpenSearchServiceConnection
-from api.services.search_interface import SearchInterface
+from api.services.opensearch.opensearch import OpenSearchServiceConnection
+from api.services.opensearch.search_interface import SearchInterface
 from opensearchpy.exceptions import OpenSearchException
 
 logger = logging.getLogger(__name__)
