+++ src/django/api/tests/test_opensearch_response_formatter.py
@@ -1,6 +1,6 @@
 import unittest
 from unittest.mock import patch, MagicMock
-from api.services.search import \
+from api.services.opensearch.search import \
     OpenSearchService, OpenSearchServiceException
