+++ src/django/api/views/v1/production_locations.py
@@ -2,11 +2,12 @@ from django.http import QueryDict
 from rest_framework import status
 from rest_framework.viewsets import ViewSet
 from rest_framework.response import Response
+
 from api.views.v1.utils import (
     serialize_params,
     handle_errors_decorator
 )
-from api.services.search import OpenSearchService
+from api.services.opensearch.search import OpenSearchService
 from api.views.v1.opensearch_query_builder.opensearch_query_builder \
     import OpenSearchQueryBuilder
 from api.views.v1.opensearch_query_builder.opensearch_query_director \
